New Authorisation
  Previously, ranger would expect to be protected by a reverse proxy
which uses mod_auth_openidc, which would provide the user's email in the
X-Remote-User header after authentication. Ranger would perform a lookup
in the database to check that this user was a member of all required
access groups.
  Now, we have two authorisation strategies: by user (as above) and by
token (provided by npg_sentry). To simultaneously support both, ranger
will expect request urls to begin with /authuser or /authtoken. Both
will make requests to npg_sentry, which performs the database lookups.
  This means that ranger no longer needs to know about user<->group
mappings, and can instead push this knowledge to npg_sentry.
  mod_auth_openidc is still required to protect the /authuser urls, but
should NOT protect the /authtoken urls.
