+++ lib/constants.js
@@ -3,9 +3,13 @@
 const TOKEN_BEARER_KEY_NAME = 'npg_sentry_token_bearer';
 const AUTH_TYPE_TOKEN = 'authtoken';
 const AUTH_TYPE_USER = 'authuser';
+const AUTH_URL_TOKEN = '/checkToken';
+const AUTH_URL_USER = '/checkUser';
 
 module.exports = {
   TOKEN_BEARER_KEY_NAME,
   AUTH_TYPE_TOKEN,
   AUTH_TYPE_USER,
+  AUTH_URL_TOKEN,
+  AUTH_URL_USER,
 };
