+++ lib/server/controller.js
@@ -23,7 +23,6 @@ const url         = require('url');
 const LOGGER      = require('winston');
 const DataMapper  = require('./mapper.js');
 const DataAccess  = require('./auth.js');
-const TokenDataAccess = require('./tokenauth.js');
 const RangerModel = require('./model.js');
 const HttpError   = require('./http/error.js');
 const trailer     = require('./http/trailer.js');
@@ -241,15 +240,9 @@ class RangerController {
       if (options.get('skipauth')) {
         cback(query);
       } else {
-        let da;
-        let authinput;
-        if (authtype === constants.AUTH_TYPE_USER) {
-          da = new DataAccess(this.db);
-          authinput = user.username;
-        } else if (authtype === constants.AUTH_TYPE_TOKEN) {
-          da = new TokenDataAccess();
-          authinput = user.token;
-        }
+        let da = new DataAccess(authtype);
+        // TODO unknown/undefined authtype should be caught in handleRequest(),
+        // is an else clause needed?
         da.once('authorised', (username) => {
           da.removeAllListeners();
           LOGGER.debug(`User ${username} is given access`);
@@ -261,7 +254,13 @@ class RangerController {
           this.errorResponse(this.response, 403,
             `Authorization failed for user '${username}'`);
         });
-        da.authorise(authinput, data.map( (d) => {return d.accessGroup;} ));
+        let identifier;
+        if (authtype === constants.AUTH_TYPE_USER) {
+          identifier = user.username;
+        } else if (authtype === constants.AUTH_TYPE_TOKEN) {
+          identifier = user.token;
+        }
+        da.authorise(identifier, data.map( (d) => {return d.accessGroup;} ));
       }
     });
     dm.getFileInfo(query, host);
@@ -526,7 +525,10 @@ class RangerController {
     let q = urlObj.query;
 
     let authmatch = path.match(new RegExp('^\/(' + constants.AUTH_TYPE_USER + '|' + constants.AUTH_TYPE_TOKEN + ')'));
-    let authtype = authmatch[1] || null;
+    let authtype = authmatch ? authmatch[1] : null;
+
+    // Finished handling authtype, so strip `/${authtype}` from url
+    path = path.replace(authmatch[0], '');
 
     this._user = {};
     if (options.get('skipauth')) {
@@ -546,9 +548,6 @@ class RangerController {
     } else {
       return this.errorResponse(this.response, 401, 'Unrecognised authtype');
     }
-
-    // Finished handling authtype, so strip `/${authtype}` from url
-    path = path.replace(authmatch[0]);
 
     let duplicateAttr = (query) => {
       return Object.keys(query).find(
